+++ lib/lightning_web/live/workflow_live/edit.ex
@@ -492,7 +492,7 @@ defmodule LightningWeb.WorkflowLive.Edit do
           |> Lightning.Repo.preload([
             :edges,
             triggers: Trigger.with_auth_methods_query(),
-            jobs: [:credential]
+            jobs: {Workflows.jobs_ordered_subquery(), [:credential]}
           ])
 
         socket |> assign_workflow(workflow) |> assign(page_title: workflow.name)