+++ CHANGELOG.md
@@ -167,6 +167,8 @@ bearing with us as we move towards our first stable Lightning release.)
   [#1348](https://github.com/OpenFn/Lightning/issues/1348)
 - Modify CompleteRun to return error changeset when run not found
   [#1393](https://github.com/OpenFn/Lightning/issues/1393)
+- Pending state on WO doesn't appear until retries are claimed
+  [#1340](https://github.com/OpenFn/Lightning/issues/1340)
 
 ## [v0.9.3] - 2023-09-27