+++ CHANGELOG.md
@@ -172,7 +172,8 @@ bearing with us as we move towards our first stable Lightning release.)
   [#1393](https://github.com/OpenFn/Lightning/issues/1393)
 - Drop invocation reasons from DB
   [#1412](https://github.com/OpenFn/Lightning/issues/1412)
-
+- Fix inconsistency in ordering of child nodes in the workflow diagram
+  [#1406](https://github.com/OpenFn/Lightning/issues/1406)
 
 ## [v0.9.3] - 2023-09-27